Move nvtx annotation utilities to a separate file
This will enable us to use them in the spilling code without circular
import issues.
